Show PNotify icons using FontAwesome4
Also:
- fix notify service filename (notification -> notify)
- remove "click to close" behaviour because the close button is now
  correctly displayed
- add tests for notify.notice and notify.error
