+++ test/client/dashboards/dashboard-image.component.spec.js
@@ -1,7 +1,7 @@
 require('../../../public/js/core/core.module.js');
 require('../../../public/js/core/constants.js');
 require('../../../public/js/core/connection.js');
-require('../../../public/js/core/notification.service');
+require('../../../public/js/core/notify.service.js');
 require('../../../public/js/core/api.js');
 require('../../../public/js/dashboards/dashboards.module.js');
 require('../../../public/js/dashboards/dashboard-image.component.js');
