+++ test/client/roles/role-edit-modal.component.spec.js
@@ -2,7 +2,7 @@ require('../../../public/js/core/core.module.js');
 require('../../../public/js/core/constants.js');
 require('../../../public/js/core/api.js');
 require('../../../public/js/core/connection.js');
-require('../../../public/js/core/notification.service');
+require('../../../public/js/core/notify.service.js');
 require('../../../public/js/core/user.service.js');
 require('../../../public/js/roles/roles.module.js');
 require('../../../public/js/roles/role-edit-modal.component.js');
