+++ views/index.ejs
@@ -16,8 +16,8 @@
         <link rel="stylesheet" href="s/font-awesome/css/font-awesome.min.css">
         <link rel="stylesheet" href="s/jsplumb/css/jsplumbtoolkit-defaults.css">
         <link rel="stylesheet" href="s/intro.js/introjs.min.css">
-        <link rel="stylesheet" href="s/pnotify-bootstrap3/PNotifyBootstrap3.css">
         <link rel="stylesheet" href="s/pnotify-core/PNotify.css">
+        <link rel="stylesheet" href="s/pnotify-bootstrap3/PNotifyBootstrap3.css">
         <link rel="stylesheet" href="s/pivottable/pivot.min.css">
         <link rel="stylesheet" href="s/subtotal/subtotal.min.css">
 
@@ -49,6 +49,7 @@
 <script src="s/angular-route/angular-route.min.js"></script>
 <script src="s/pnotify-core/PNotify.js"></script>
 <script src="s/pnotify-bootstrap3/PNotifyBootstrap3.js"></script>
+<script src="s/pnotify-fontawesome4/PNotifyFontAwesome4.js"></script>
 <script src="s/moment/moment-with-locales.min.js"></script>
 <script src="s/angularjs-bootstrap-datetimepicker/js/datetimepicker.js"></script>
 <script src="s/angularjs-bootstrap-datetimepicker/js/datetimepicker.templates.js"></script>
@@ -83,7 +84,7 @@
 <script type="text/javascript" src="js/core/local-storage.js"></script>
 <script type="text/javascript" src="js/core/user.service.js"></script>
 <script type="text/javascript" src="js/core/uuid.service.js"></script>
-<script type="text/javascript" src="js/core/notification.service.js"></script>
+<script type="text/javascript" src="js/core/notify.service.js"></script>
 <script type="text/javascript" src="js/core/pager.component.js"></script>
 <script type="text/javascript" src="js/core/language.js"></script>
 <script type="text/javascript" src="js/core/language-select.component.js"></script>
