+++ views/login.ejs
@@ -22,6 +22,7 @@
     <script src="s/angular-gettext/angular-gettext.min.js"></script>
     <script src="s/pnotify-core/PNotify.js"></script>
     <script src="s/pnotify-bootstrap3/PNotifyBootstrap3.js"></script>
+    <script src="s/pnotify-fontawesome4/PNotifyFontAwesome4.js"></script>
     <script src="s/moment/moment-with-locales.min.js"></script>
     <script src="s/numeral/numeral.min.js"></script>
     <script src="s/angular-ui-bootstrap/ui-bootstrap-tpls.js"></script>
